+++ CHANGELOG.md
@@ -5,6 +5,6 @@
 ### Added
 
 * `StateCurrentCmd` to discover the current state.
-* `StateDesiredCmd` to discover the desired state.
+* `StateDesiredDiscoverCmd` to discover the desired state.
 * `DiffCmd` to compute the difference between the current and desired states.
 * `EnsureCmd` to transform the current state into the desired state.